requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
boto3==1.38.19
botocore==1.38.19
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
distlib==0.3.9
distro==1.9.0
filelock==3.18.0
Flask==3.1.1
fla***REMOVED***cors==6.0.0
Fla***REMOVED***SQLAlchemy==3.1.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
lxml==5.4.0
MarkupSafe==3.0.2
openai==1.81.0
packaging==25.0
pbr==6.1.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
PyMuPDF==1.25.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
s3transfer==0.12.0
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stevedore==5.4.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.31.2
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
Werkzeug==3.1.3
